chore(example): separted examples
separated and simplified examples for each adapter, setup commands to run independently
