+++ examples/basic/package.json
@@ -4,7 +4,9 @@
   "description": "",
   "private": true,
   "scripts": {
-    "start": "tsx --env-file=.env ./src/main.ts",
+    "start": "run-s start:*",
+    "start:vertex": "tsx --env-file=.env ./src/vertex.ts",
+    "start:openai": "tsx --env-file=.env ./src/openai.ts",
     "type-check": "tsc --noEmit"
   },
   "keywords": [],
@@ -13,6 +15,7 @@
   "type": "module",
   "devDependencies": {
     "@types/node": "18.19.32",
+    "npm-run-all": "^4.1.5",
     "tsx": "^4.11.2",
     "typescript": "5.4.5"
   },
