+++ examples/basic/src/vertex.ts
@@ -1,9 +1,7 @@
 import { Kurt } from "@formula-monks/kurt"
 import { KurtVertexAI } from "@formula-monks/kurt-vertex-ai"
-import { KurtOpenAI } from "@formula-monks/kurt-open-ai"
-import { OpenAI } from "openai"
 import { VertexAI } from "@google-cloud/vertexai"
-import { testRun } from "./test-run"
+import { z } from "zod"
 
 const vertexAdapter = new KurtVertexAI({
   vertexAI: new VertexAI({
@@ -13,13 +11,15 @@ const vertexAdapter = new KurtVertexAI({
   model: "gemini-1.0-pro", // or any other supported model
 })
 
-const openAIAdapter = new KurtOpenAI({
-  openAI: new OpenAI(),
-  model: "gpt-3.5-turbo-0125",
+const kurt = new Kurt(vertexAdapter)
+
+const stream = kurt.generateStructuredData({
+  // or any other generation method
+  prompt: "Say hello!",
+  schema: z.object({
+    say: z.string().describe("A single word to say"),
+  }),
 })
 
-console.log("Testing Vertex Adapter")
-await testRun(new Kurt(vertexAdapter))
-
-console.log("Testing OpenAI Adapter")
-await testRun(new Kurt(openAIAdapter))
+const { data } = await stream.result
+console.log(data) // { say: "hello" }
